Device records: prepare for gap queries
What is this feature for? You'll be able to ask for last 1000 record without gap and you'll get 1000 records. If you ask for 1000 records with gap=2 it will still return 1000 records but it takes into account 2000 records and it will return every second record. Therefore you can see to the deeper history but still with the same amount of returned data.